Quarantine implementation qgep_import
Insert/update of qgep_import.vw_manhole inserts the data into the qgep_import.quarantine_manhole.
An insert/update of qgep_import.quarantine_manhole tries an update on the wastewater_structure data, cover and reach/reachpoint.
An update of the reach/reachpoint is done automatically when there is exactly 1 inlet and 1 outlet in the incoming data and the existing data.
If not, the user has to enter them manually and change inlet_okay/outlet_okay to true, that the entry in the quarantine is deleted.

Photo saving is not working yet!

Including tests for the mentioned functionalities.